Changed habitability to double and added seppo's magic to setHabitability

+++ Codename Frozen Omega/src/CodenameFrozenOmega/Animal.java
@@ -2,7 +2,7 @@ package CodenameFrozenOmega;
 
 public class Animal extends Organism {
 	
-	protected Animal(int preferredTemperature, int id, int location, double lifespan, int habitability,
+	protected Animal(int preferredTemperature, int id, int location, double lifespan, double habitability,
 						double age, String species, int foodChainPosition, int birthTick, boolean healthy) {
 		super(preferredTemperature, id, location, habitability, species, age, birthTick, lifespan, healthy);
